+++ NEWS.md
@@ -1,3 +1,5 @@
+# tabnet (development version)
+
 # tabnet 0.3.0
 
 - Added an `update` method for tabnet models to allow the correct usage of `finalize_workflow` (#60).