+++ NEWS.md
@@ -6,9 +6,9 @@
 * Improve unsupervised vignette content. (#67)
 * `tabnet_pretrain()` now allows missing values in predictors. (#68)
 * `tabnet_explain()` now works for `tabnet_pretrain` models. (#68)
-* Add  vignette for "pretrain from dataset with missing-values". (#68)
+* Allow missing-values values in predictor for unsupervised training. (#68)
 * Improve performance of `random_obfuscator()` torch_nn module. (#68)
-* add support for early stopping (#69)
+* Add support for early stopping (#69)
 
 # tabnet 0.3.0