+++ NEWS.md
@@ -1,5 +1,7 @@
 # tabnet (development version)
 
+# tabnet 0.4.0
+
 ## New features
 
 * Add explicit legend in `autoplot.tabnet_fit()` (#67)